+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to Calendar Map Filter (CMF) will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.9] - 2025-10-6
+
+### Added
+
+- Support for user to jump to current location on the map by clicking icon in bottom right, aka GeolocateControl, 
+  using browser geolocation API. User must give permission.
+
 ## [0.3.8] - 2025-10-6
 
 ### Added